Conflicts:
	src/FBDialogs.m
	src/FBLoginView.m
	src/FBUserSettingsViewController.m
	src/FBViewController.m

+++ src/FBRequestBody.m
@@ -6,7 +6,7 @@
  * You may obtain a copy of the License at
  *
  *    http://www.apache.org/licenses/LICENSE-2.0
- 
+ *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -15,6 +15,7 @@
  */
 
 #import "FBRequestBody.h"
+
 #import "FBSettings+Internal.h"
 
 static NSString *kStringBoundary = @"3i2ndDfv2rTHiSisAbouNdArYfORhtTPEefj3q2f";
@@ -88,7 +89,7 @@ static NSString *kStringBoundary = @"3i2ndDfv2rTHiSisAbouNdArYfORhtTPEefj3q2f";
     NSData *data = UIImageJPEGRepresentation(image, [FBSettings defaultJPEGCompressionQuality]);
     [self.mutableData appendData:data];
     [self appendRecordBoundary];
-    [logger appendFormat:@"\n    %@:\t<Image - %d kB>", key, [data length] / 1024];
+    [logger appendFormat:@"\n    %@:\t<Image - %lu kB>", key, (unsigned long)([data length] / 1024)];
 }
 
 - (void)appendWithKey:(NSString *)key
@@ -101,7 +102,7 @@ static NSString *kStringBoundary = @"3i2ndDfv2rTHiSisAbouNdArYfORhtTPEefj3q2f";
     [self appendUTF8:@"Content-Type: content/unknown\r\n\r\n"];
     [self.mutableData appendData:data];
     [self appendRecordBoundary];
-    [logger appendFormat:@"\n    %@:\t<Data - %d kB>", key, [data length] / 1024];
+    [logger appendFormat:@"\n    %@:\t<Data - %lu kB>", key, (unsigned long)([data length] / 1024)];
 }
 
 - (NSData *)data
